backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.4
spacy==3.7.4
sentence-transformers==2.6.1
scikit-learn==1.4.1.post1
sqlalchemy==2.0.27
aiosqlite==0.20.0
greenlet==3.0.3
pydantic==2.6.1
pydantic-settings==2.2.0
python-dotenv==1.0.1
numpy==1.26.4
aiofiles==23.2.1
httpx==0.26.0
reportlab==4.1.0
easyocr==1.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymupdf==1.24.1
Pillow==10.3.0
